Working on a lot of stuff
-Camera: Changed camera back to being controlled by player.
-Player controllable, still sort of broken.
-Collisions work great.
-I like the way that object rects are moved by camera through a
delegate. going to try to make all objects behave like this.
-Moved keyboard controls back to player
-Attempting to delete InteractiveCharacter class
-Created a Levelhandler class, idealy when this is finished rather than
having any level changing logic in Game1.cs, it will all be in
LevelHandler, and we can just call LevelHandler.Update()

Unprofessional note:
It may seem shitty to change coordinate system back to doing math on
virtual coordinates to get draw coordinates, but it will be better in
the end i promise
-